requirements.txt
pandas
geopandas 
shapely 
requests
pystac-client
rasterio
planetary-computer
python-dotenv
tqdm
matplotlib
pyqt5